Create README.md
Hangman is a word game, where the goal is to find the missing word or words. Asterisks will appear in place of the missing word.
Use the keyboard to guess a letter. If you choose a letter exists in your answer, will appear the corresponding message (" Sorry, you already used that letter!
Try another, one!").
Every time you guess a letter wrong you loose a life and the hangman begins to appear, piece by piece.
At the end of the game, the corresponding message and the number of attempts will be displayed.